% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsearchr.R
\name{cdscheck}
\alias{cdscheck}
\title{Check for the status of a submitted search on NCBI's CD-SEARCH.}
\usage{
cdscheck(cdsid_check = NULL, tdata_check = NULL, cdsurl_check = NULL,
maxreps = NULL, sleep_duration = NULL)
}
\arguments{
\item{cdsid_check}{(character string, mandatory) the \code{cdsid} returned by
the NCBI CD-SEARCH server upon successful submission of a search request.}

\item{tdata_check}{(character string, mandatory) the format in which the
target data should be retrieved.}

\item{cdsurl_check}{(character string, mandatory) URL at which the CD-SEARCH
server can be accessed.}

\item{maxreps}{(integer, mandatory) maximum number of attempts to be made to
query the server on the status of the search before giving up.}

\item{sleep_duration}{(integer, mandatory) maximum duration in seconds between
successive queries to the server.}
}
\value{
1 if the search terminated successfully, otherwise 0.
}
\description{
Check on the status of a submitted search associated with the given \code{cdsid}
(@seealso \code{\link{cdssubmit}}). The values for all the parameters below
are inherited from \code{\link{cdsearchr}} (@seealso \code{\link[=cdsearchr]{cdsearchr()}} for some details).
This function is also responsible for interpreting (and printing to the
terminal) the CD-SEARCH-specific error-codes.
}
