% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsearchr.R
\name{cdsrestodf}
\alias{cdsrestodf}
\title{Extract CD-SEARCH results as data.frame.}
\usage{
cdsrestodf(content, rowstoskip = 7, sep = "\t")
}
\arguments{
\item{content}{(\code{\link{httr}} response object, mandatory) \code{\link{httr}} response object
containing the annotation results from the CD-SEARCH server.}

\item{rowstoskip}{(numeric, mandatory) the response contains header rows for
the results table. These need to be skipped. This indicates how many such rows
there are.}

\item{sep}{(character, mandatory) separator symbol delineating successive rows
of the results table.}
}
\value{
A \code{\link[base:data.frame]{base::data.frame}}.
}
\description{
\code{\link{cdsearchr}} internal function that takes the HTML-formatted results from
\code{\link{cdsretrieve}} and converts it into a \code{\link[base:data.frame]{base::data.frame}} using
\code{\link[data.table:fread]{data.table::fread}}.
}
