% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdsearchr.R
\name{cdsretrieve}
\alias{cdsretrieve}
\title{Retrieve annotation results from CD-SEARCH.}
\usage{
cdsretrieve(cdsid, tdata, qdefl, cddefl, dmode, cdsurl)
}
\arguments{
\item{cdsid}{(character string, mandatory) the \code{cdsid} returned by
the NCBI CD-SEARCH server upon successful submission of a search request.}

\item{tdata}{(character string, mandatory) the format in which the
target data should be retrieved.}

\item{qdefl}{(binary, optional) should query titles be included in the
results? (Set to TRUE by default.)}

\item{cddefl}{(binary, optional) should domain titles be included in the
results? (Set to FALSE by default.)}

\item{dmode}{(character string, optional) which data mode must be used for
the results.}

\item{cdsurl}{(character string, mandatory) URL at which the CD-SEARCH
server can be accessed.}
}
\value{
A \code{\link{httr}} response object containing a (HTML-formatted) text table of
annotation results from CD-SEARCH.
}
\description{
cdsearchr() internal function that retrieves the (HTML-formatted)
tabular data containing the annotations for a given \code{cdsid}
(@seealso \code{\link[=cdssubmit]{cdssubmit()}}). All parameters are inherited from
\code{\link[=cdsearchr]{cdsearchr()}}.
}
