% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastodf.R
\name{fasdirdf}
\alias{fasdirdf}
\title{Read in all FASTA files in a directory as a data.frame.}
\usage{
fasdirdf(path = NULL, pat = NULL, seqtype = c("DNA", "AA"), incl_filepath = TRUE)
}
\arguments{
\item{path}{(character string, mandatory) the path to the directory containing
the input FASTA files.}

\item{pat}{(character string, optional) a regex string as used by \link[base]{list.files}
to specify which file/directory names should be returned.}

\item{seqtype}{(character string, optional) the type of sequence being read in;
DNA ("DNA") or amino acid ("AA"). Defaults to "DNA".}

\item{incl_filepath}{(logical, optional) should the path to the file being read
be included in the \code{data.frame}. Irrespective of whether the user sets this
to TRUE or FALSE, a column (\code{filename}) will be included in the output
\code{data.frame} to keep the number of columns consistent.}
}
\value{
a \link[base]{data.frame} with the following columns: \code{seqname},
\code{seq}, and \code{filename}. If \code{incl_filepath} is set to TRUE, then
\code{filename} will include the full path to the input file. It will be set to
\code{NA} otherwise.
}
\description{
Wrapper around \link{fastodf} that reads in all FASTA files in a
directory into a single \link[base]{data.frame}. Uses \link{list_files}
internally also. At present, \code{fasdirf} requires that all files being read in
are of the same type (DNA or amino acid).
}
\examples{
\dontrun{
#Input data
inpath <- dirname(system.file("extdata", "cdsearchr_testdata.fasta",
                              package = "seqvisr", mustWork = TRUE))
#Reading in some sample amino acid sequences
fasdirdf(path = inpath, seqtype = "AA", pat = "*_testdata.fasta")
}

}
