% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastodf.R
\name{fastodf}
\alias{fastodf}
\title{Read in a FASTA file as a data.frame.}
\usage{
fastodf(path = NULL, seqtype = c("DNA", "AA"), incl_filepath = TRUE)
}
\arguments{
\item{path}{(character string, mandatory) the path to the input FASTA file.}

\item{seqtype}{(character string, optional) the type of sequence being read in;
DNA ("DNA") or amino acid ("AA"). Defaults to "DNA".}

\item{incl_filepath}{(logical, optional) should the path to the file being read
be included in the \code{data.frame}. Irrespective of whether the user sets this
to TRUE or FALSE, a column (\code{filename}) will be included in the output
\code{data.frame} to keep the number of columns consistent.}
}
\value{
a \link[base]{data.frame} with the following columns: \code{seqname},
\code{seq}, and \code{filename}. If \code{incl_filepath} is set to TRUE, then
\code{filename} will include the full path to the input file. It will be set to
\code{NA} otherwise.
}
\description{
Utility function that uses \link[seqinr]{read.fasta} to load a FASTA file as a
\link[base]{data.frame}.
}
\examples{
\dontrun{
#Input data
inpath <- system.file("extdata", "cdsearchr_testdata.fasta", package = "seqvisr", mustWork = TRUE)
#Reading in some sample amino acid sequences
fastodf(path = inpath, seqtype = "AA")
}

}
