% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqvisr-package.R
\docType{package}
\name{seqvisr-package}
\alias{seqvisr}
\alias{seqvisr-package}
\title{seqvisr: Biological Sequence Visualization and Auxiliary Functions in R}
\description{
Enables visualization of biological sequences (and manipulation thereof, among other things) in R.
}
\details{
R functions for visualization of biological sequences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vragh/seqvisr}
  \item Report bugs at \url{https://github.com/vragh/seqvisr/issues}
}

}
\author{
\strong{Maintainer}: Venket Raghavan \email{vraghvragh@mailfence.com}

}
\keyword{"multiple}
\keyword{"protein}
\keyword{"sequence}
\keyword{alignment"}
\keyword{domains"}
\keyword{sequence}
\keyword{visualization"}
