% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqvisr-package.R
\docType{package}
\name{seqvisr-package}
\alias{seqvisr}
\alias{seqvisr-package}
\title{seqvisr: Biological sequence visualization functions in R}
\description{
seqvisr enables visualization of biological sequences (and manipulation thereof) in R. It currently comprises of a single function msavisr that can be used to visualize multiple sequence alignments. Additional functions (and functionality) to be added in the future.
}
\details{
The only function available from seqvisr at the moment is \code{\link[=msavisr]{msavisr()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vragh/seqvisr}
  \item Report bugs at \url{https://github.com/vragh/seqvisr/issues}
}

}
\author{
\strong{Maintainer}: Venket Raghavan \email{vraghvragh@mailfence.com}

}
\keyword{internal}
